rename: view 폴더 생성 및 관련 파일 경로 수정 (InputView.js, OutputView.js, Messages.js)
